优化自动执行逻辑
