+++ registry/lib/components/utils/check-in-center/Widget.vue
@@ -16,15 +16,6 @@
 import { DefaultWidget } from '@/ui'
 import { CheckInItem, checkInItems } from './check-in-item'
 
-const shouldAutoRun = (item: CheckInItem): boolean => {
-  if (!item.autoRun) {
-    return false
-  }
-  const last = new Date(item.lastRun).toDateString()
-  const now = new Date().toDateString()
-  return last !== now
-}
-
 export default Vue.extend({
   components: {
     DefaultWidget,
@@ -34,24 +25,7 @@ export default Vue.extend({
       items: checkInItems,
     }
   },
-  mounted() {
-    this.autoCheckIn()
-  },
   methods: {
-    async autoCheckIn() {
-      for (const item of checkInItems) {
-        if (shouldAutoRun(item)) {
-          try {
-            this.$set(item, 'disabled', true)
-            await item.action(document.createElement('div'), new MouseEvent('click'))
-          } catch (error) {
-            console.error(`${item.displayName} 自动执行请求失败`, error)
-          } finally {
-            item.disabled = false
-          }
-        }
-      }
-    },
     async runItemAction(item: CheckInItem, event: MouseEvent) {
       try {
         // 一开始可能是 undefined
